backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.10
flask-migrate==4.0.5
Flask-Caching==2.1.0
redis==5.0.3 